+++ $notes/changelog.md
@@ -2,6 +2,7 @@
 ## Added botany pots
 - Added support for let's do cooking mod crops
 - Added BotanyPotsTiers for ability to upgrade pots for better efficiency.
+- Added support for PamHC2Trees saplings producing fruits in pots
 
 # 3.6: sitting/easy farm
 ## Added a suit of mods for sitting support